cambios en los servicios y repositorios

+++ src/main/java/co/edu/uniquindio/bookyourstay/modelo/BilleteraVirtual.java
@@ -9,6 +9,6 @@ import lombok.Setter;
 @Getter
 public class BilleteraVirtual {
 
-    private int saldo;
+    private float saldo;
 
 }
